Add encrypted kv_store TTL regression test
