+++ .gitignore
@@ -1,0 +1,7 @@
+# Build artifacts generated by make -C Test and module builds
+**/objs/
+**/*.a
+Test/libft_tests
+test_*.txt
+
+test_failures.log
