+++ Storage/Makefile
@@ -1,7 +1,9 @@
 TARGET         := storage.a
 DEBUG_TARGET   := storage_debug.a
 
-SRCS := storage_kv_store.cpp
+SRCS := storage_kv_store_constructor.cpp \
+        storage_kv_store_encryption.cpp \
+        storage_kv_store_operations.cpp
 
 HEADERS := kv_store.hpp
